Made some minor changes implemented actual db url

+++ source/database/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS products (
     productId SERIAL PRIMARY KEY,
     productName VARCHAR(255),
-    productAmount NUMBER(20) NOT NULL
+    productAmount BIGINT NOT NULL
 );
